requirements.txt
beautifulsoup4 @ file:///home/conda/feedstock_root/build_artifacts/beautifulsoup4_1601745390275/work
brotlipy==0.7.0
certifi==2020.6.20
cffi @ file:///Users/runner/miniforge3/conda-bld/cffi_1602537233343/work
chardet @ file:///Users/runner/miniforge3/conda-bld/chardet_1602255311078/work
cryptography @ file:///Users/runner/miniforge3/conda-bld/cryptography_1604179238896/work
idna @ file:///home/conda/feedstock_root/build_artifacts/idna_1593328102638/work
pycparser @ file:///home/conda/feedstock_root/build_artifacts/pycparser_1593275161868/work
pyOpenSSL==19.1.0
PySocks @ file:///Users/runner/miniforge3/conda-bld/pysocks_1602326930468/work
requests @ file:///home/conda/feedstock_root/build_artifacts/requests_1592425495151/work
six @ file:///home/conda/feedstock_root/build_artifacts/six_1590081179328/work
soupsieve @ file:///home/conda/feedstock_root/build_artifacts/soupsieve_1597680516047/work
urllib3 @ file:///home/conda/feedstock_root/build_artifacts/urllib3_1603125704209/work

requests==2.24.0
pandas==1.1.1
beautifulsoup4==4.9.3
mysql-connector-python==8.0.18